refactor: image path관련 함수 위치 변경 및 작성시간 hydration 오류 수정
