+++ app/lib/utils/supabase/images.ts
@@ -34,32 +34,3 @@ export async function removeImages(folder: string, imagesToRemain?: string[]) {
     .from("patda-images")
     .remove(imagePathesToDelete);
 }
-
-export function getTempFolderPath(session: Session, postId?: number) {
-  const userNameKey = encodeURIComponent(session.user?.name ?? "").replace(
-    /[^a-zA-Z0-9]/g,
-    ""
-  );
-  const userId = session.user?.id ?? "";
-
-  const postKey = postId !== undefined ? `${postId}/` : "";
-
-  return `temp/${postKey}${userNameKey}${userId.slice(0, 3)}`;
-}
-
-interface GetImagePathProps {
-  session?: Session;
-  postId?: number;
-  commentId?: string;
-}
-
-export function getImagePath({
-  session,
-  postId,
-  commentId,
-}: GetImagePathProps) {
-  if (session) return getTempFolderPath(session, postId);
-  if (postId !== undefined) return `post/${postId}`;
-  if (commentId !== undefined) return `comment/${commentId}`;
-  return "";
-}
